Optimize and fix class template lookup
With this patch, we choose one template instantiation as
a representative for each template name when building
the (manual) index.

As a result, when we find a template in the index, we
only build AST for one instantiation rather than for
all the instantiations. This greatly speeds up the first
lookup of each template name (subsequent lookups have
been already fast because the ASTs are already built).
In Unreal engine, we see the first evaluation involving
the TArray template go from 4s to 0.8s.

This patch also fixes template lookup in namespaces.