Clear children of ValueObject on value update
Children of ValueObject automatically update themselves when they detect
the state of the process has changed, which typically happens when the
parent value is updated. However, if in case of updating
ValueObjectConstResult the process state is unchanged and the children
remain stale.

Explicitly clear the children upon the parent update, so that they're
re-calculated afterwards.

Differential Revision: https://reviews.llvm.org/D105470

NOTE: The upstream patch is not submitted as there is an ongoing
discussion. However I believe the patch is "good enough" for our use
case, so I want to proceed with it as is and replace with a better
solution later.