Extract the qualified name from the DIE
Getting the qualified name from the Type is expensive, because it
requires building a forward declaration of the type. In the extreme case
of having thousands of elements in the TypeMap the initial lookup takes
~minutes.

This patch speeds things up for a particular case of using DWARF symbol
files. The improvement for the initial lookup for Unreal5 ShooterGame:

```
# Before
(lldb) script timeit(lldb.target.FindTypes("EAutoReceiveInput::Type"))
83.72795928385828
# After
(lldb) script timeit(lldb.target.FindTypes("EAutoReceiveInput::Type"))
7.014563322067261
```