[AArch64] Skip over the shadow store for ARM64EC entry thunk variadic calls
When in an entry thunk the x64 SP is passed in x4 but this cannot be directly
passed through since x64 varargs calls have a 32 byte shadow store at SP
followed by the in-stack parameters. Arm64EC varargs calls on the other hand
expect x4 to point to the first in-stack parameter.